chore(routes): reorganized destination-related requests into separate folder for better organization
