+++ app/Http/Requests/Destination/IndexDestinationRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Destination;
 
 use Illuminate\Foundation\Http\FormRequest;
